# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ANPE CENTRE NORD API",
        "description": "Bienvenue dans la documentation de l'API de l'ANPE CENTRE NORD.<br>Cette API fournit un accès sécurisé aux fonctionnalités du backend, <br>permettant la gestion efficace des ressources liées à l'emploi, aux utilisateurs et aux données associées.",
        "contact": {
            "name": "SYMBOL SARL",
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur test de l'API"
        }
    ],
    "paths": {
...
        "/api/profile-experiences": {
            "get": {
                "tags": [
                    "Profile Expérience"
                ],
                "summary": "Liste des expériences",
                "description": "Retourne la liste des expériences",
                "operationId": "33ec9d1bdc5f7baaf43750bdd4efa33d",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Profile Expérience"
                ],
                "summary": "Création d'une expérience",
                "description": "Crée une nouvelle expérience et retourne l'expérience créée",
                "operationId": "e36d38f8538d21f61f39891a4cf99957",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label"
                                ],
                                "properties": {
                                    "titre": {
                                        "type": "string",
                                        "example": "Mise en place d'un plugin moodle"
                                    },
                                    "entreprise": {
                                        "type": "string",
                                        "example": "SYMBOL-SARL"
                                    },
                                    "date_debut": {
                                        "type": "string",
                                        "example": "01/01/2024"
                                    },
                                    "date_fin": {
                                        "type": "string",
                                        "example": "01/04/2024"
                                    },
                                    "profile": {
                                        "type": "string",
                                        "example": "Slug du profile"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Lorem ipsum lorem"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Expérience créée avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies ne sont pas valides"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile-experiences/{slug}": {
            "get": {
                "tags": [
                    "Profile Expérience"
                ],
                "summary": "Récupère une expérience par son slug",
                "description": "Retourne une expérience",
                "operationId": "1670b9cc405e602cc276d6d13b9c056d",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug de l'expérience à récupérer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile Expérience"
                ],
                "summary": "Modification d'une expérience",
                "description": "Modifie une expérience et retourne l'expérience modifiée",
                "operationId": "a1773b6ada99720eee856f05559181b1",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug de l'expérience à modifiée",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label"
                                ],
                                "properties": {
                                    "titre": {
                                        "type": "string",
                                        "example": "Mise en place d'un plugin moodle"
                                    },
                                    "entreprise": {
                                        "type": "string",
                                        "example": "SYMBOL-SARL"
                                    },
                                    "date_debut": {
                                        "type": "string",
                                        "example": "01/01/2024"
                                    },
                                    "date_fin": {
                                        "type": "string",
                                        "example": "01/04/2024"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Lorem ipsum lorem"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Expérience modifiée avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slug validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Expérience non trouvée"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies ne sont pas valides"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profile Expérience"
                ],
                "summary": "Supprime une expérience par son slug",
                "description": "Retourne l'expérience supprimée",
                "operationId": "573c2d752db2dcbd1383077560647114",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug de l'expérience à supprimer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Expérience supprimée avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slug validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Expérience non trouvée"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}